CRUD operation. Insert operation remained (user)
