+++ src/components/pages/EditData/EditUser.jsx
@@ -1,0 +1,34 @@
+import React from "react";
+import { useQuery } from "react-query";
+import { useParams, useNavigate } from "react-router";
+import { fetchSingleUser } from "../../../hooks/users/useFindUser";
+import UserForm from "../../elements/UserForm/UserForm";
+// Pages
+import Aside from "../../layouts/Aside/Aside";
+import Profile from "../Profile/Profile";
+
+const EditData = () => {
+	const navigate = useNavigate();
+	const params = useParams();
+
+	const { data: fetchUser, isLoading, refetch } = useQuery(["edit-user", params.id], fetchSingleUser);
+
+	if (isLoading) {
+		return "Load...";
+	}
+
+	if (fetchUser?.data.length === 0) {
+		setTimeout(() => navigate("/dashboard"), 3000);
+		return <h1>Error has occured. You'll be redirect in 3 seconds.</h1>;
+	}
+	return (
+		<div className="admin-dashboard d-flex">
+			<Aside />
+			<div className="container-fluid">
+				<UserForm fetchUser={fetchUser.data[0]} isLoading={isLoading} refetch={refetch} isAdmin={true} />
+			</div>
+		</div>
+	);
+};
+
+export default EditData;
