+++ src/components/pages/Profile/Profile.js
@@ -2,194 +2,12 @@ import React, { useEffect, useState } from "react";
 // Css
 import "./Profile.css";
 // Images
-import noImage from "../../../assets/no-user-image.png";
-import { AiOutlineMail } from "react-icons/ai";
-import { FcBusiness } from "react-icons/fc";
-
 import useLoggedUser from "../../../hooks/users/useLoggedUser";
-import { useUpdateUser } from "../../../hooks/users/useUpdateUser";
-
-import { useRemoveImage } from "../../../hooks/images/useRemoveImage";
-import { useUploadPhoto } from "../../../hooks/useUploadPhoto";
-import { useQuery } from "react-query";
-
-import QuantoxSpinner from "../../elements/QuantoxSpinner/QuantoxSpinner";
-import { upload } from "@testing-library/user-event/dist/upload";
-import instance from "../../../config/config";
+import UserForm from "../../elements/UserForm/UserForm";
 
 const Profile = () => {
 	const { data: fetchUser, isLoading, refetch } = useLoggedUser();
-	const { mutate: updateUser } = useUpdateUser();
-	const { mutate: removeImage } = useRemoveImage();
-	const { mutate: uploadPhoto } = useUploadPhoto();
-
-	//Mozda bi trebalo da u useQuery ili useLoggedUser da odradim async/await da bi se dobio odgovor
-	const { data: imageObj, refetch: uploadRefetch } = useQuery("upload-info", () => {}, { cacheTime: 0, retry: true, refetchOnMount: false, refetchOnWindowFocus: false });
-
-	// /^[a-zžšđčćA-ZŽŠĐČĆ!@_\d]{2,30}$/
-	const usernameRegEx = /^([a-zžšđčćA-ZŽŠĐČĆ!@_]{2,30})+(\d)*/;
-	const passwordRegEx = /^[A-ZŽŠĐČĆa-zžšđčć\d_#!@$%^*{}]{5,30}$/;
-
-	const userData = fetchUser?.data;
-	const { id, username, email, role, profileImage } = {
-		id: userData?.id,
-		username: userData?.username,
-		email: userData?.email,
-		role: userData?.role?.name,
-		profileImage: userData?.profileImage,
-	};
-
-	// Bez ovog userName nece da mi radi onChange...
-	const [userName] = useState(undefined);
-	const [editView, setEditView] = useState(false);
-	const [message, setMessage] = useState(null);
-	const [newUserData, setNewUserData] = useState({
-		id: id,
-		username: undefined,
-		profileImage: null,
-		password: null,
-	});
-
-	// useEffect(() => {
-	// 	setUserName(username);
-	// }, [username]);
-
-	useEffect(() => {
-		setNewUserData({ ...newUserData, id: id });
-	}, [id]);
-
-	const handleUpdate = async () => {
-		setNewUserData({ ...newUserData, id });
-
-		if (newUserData.username === "" || newUserData.username === undefined || newUserData.username === null) {
-			setNewUserData({ ...newUserData, username: undefined });
-		} else if (!usernameRegEx.test(newUserData.username)) {
-			return setMessage("Username is not in valid format.");
-		} else {
-			setNewUserData({ ...newUserData, username: userName });
-		}
-
-		if (newUserData.password === "" || newUserData.password === null) {
-			setNewUserData({ ...newUserData, password: null });
-		} else if (!passwordRegEx.test(newUserData.password)) {
-			return setMessage("Password is not in valid format.");
-		}
-
-		let sendingData = { id };
-		/* 
-			Trebaju mi samo svojstva koja dolaze iz inputa. Ako svojstva prodju provere iznad, 
-			upisuju se u promenljivu data i salju se na UPDATE metodom PUT 
-		*/
-		for (let property in newUserData) {
-			if (newUserData[property] && property !== "id") {
-				sendingData = { ...sendingData, [property]: newUserData[property] };
-			}
-			if (newUserData.profileImage !== null && property === "profileImage") {
-				//Stalno ce ulazti ovde....
-
-				if (newUserData.profileImage !== null) {
-					//Prosao
-					// await uploadPhoto(newUserData.profileImage);
-					await instance.post("/api/upload", newUserData.profileImage).then((res) => {
-						console.log(res.data[0].id);
-						const imageId = res.data[0].id;
-						sendingData = { ...sendingData, profileImage: imageId };
-					});
-				}
-				//Nije prosao jer je imageObj[0] undefiend
-			}
-		}
-		console.log(sendingData);
-		updateUser(sendingData);
-		setTimeout(() => refetch(), 500);
-		setMessage(null);
-		handleView();
-
-		setNewUserData({ ...newUserData, profileImage: null, password: null });
-	};
-
-	const handleView = () => {
-		if (message !== null) {
-			setMessage(null);
-		}
-		setEditView(!editView);
-	};
-
-	const handleFileChange = (file) => {
-		const fileUploaded = file.target.files[0];
-		const imageTypes = ["image/jpeg", "image/png", "image/gif"];
-
-		//Check if someone want to upload wrong format
-		if (!imageTypes.some((type) => fileUploaded.type === type)) {
-			return setMessage("You cannot upload video or other document, only picture");
-		}
-
-		const imageData = new FormData();
-		imageData.append("files", fileUploaded);
-		setNewUserData({ ...newUserData, profileImage: imageData });
-	};
-
-	const handleRemoveImage = () => {
-		removeImage(profileImage.id);
-		setTimeout(() => refetch(), 500);
-		handleView();
-	};
-
-	return (
-		<div className="profile container container-project mt-5">
-			<div className="row">
-				<div className="col-sm-12 col-md-4">
-					<div className="card profile-image-container p-3 d-flex align-items-center">
-						{isLoading ? "Loading..." : null}
-						<div className="profile-image-box d-flex justify-content-center align-items-center">{profileImage !== null ? <img src={process.env.REACT_APP_BASEURL + profileImage?.url} alt={profileImage?.alternativeText} width={200} className={"img-fluid"} /> : <img src={noImage} alt="User has no pic" width={200} />}</div>
-					</div>
-				</div>
-				<div className="col-sm-12 col-md-8">
-					<div className="card mt-4 mt-md-0">
-						<div className="card-header bg-green">
-							<h1 className="text-muted">Data</h1>
-						</div>
-						<div className="card-body p-4 d-flex flex-column gap-3">
-							<div className="username d-flex flex-column gap-1">
-								<small className="text-danger">{message}</small>
-								{editView ? <input type="text" className="form-control" value={userName} onChange={(e) => setNewUserData({ ...newUserData, username: e.target.value })} placeholder="Enter a new username..." /> : <p className="h3 text-uppercase">{username}</p>}
-								{editView ? <input type="password" placeholder="New password..." className="form-control" onChange={(e) => setNewUserData({ ...newUserData, password: e.target.value })} /> : null}
-								{editView ? <input type="file" className="form-control" onChange={(e) => handleFileChange(e)} /> : null}
-							</div>
-							<div className="email d-flex align-items-center gap-2">
-								<AiOutlineMail />
-								<p className="text-muted">{email}</p>
-							</div>
-							<div className="role d-flex align-items-center gap-2">
-								<FcBusiness />
-								<p className="text-muted">{role}</p>
-							</div>
-							<div className="button-container">
-								<input type="button" value="Edit ⚙" className="btn btn-success me-3" onClick={handleView} />
-								{editView ? <input type="button" value="Update" className="btn btn-primary" onClick={handleUpdate} /> : ""}
-
-								{profileImage && editView ? <input type="button" value="Remove picture" className="btn btn-danger float-end" onClick={handleRemoveImage} /> : null}
-							</div>
-						</div>
-					</div>
-				</div>
-			</div>
-			<div className="row">
-				<div className="col-sm-12 col-md-4">
-					<div className="card mt-4 mt-md-0">
-						<div className="card-header">
-							<h1 className="h2 text-center text-muted">Active projects</h1>
-						</div>
-						<div className="card-body">
-							<p className="h1 text-center">{fetchUser?.data.projects?.length}</p>
-						</div>
-					</div>
-				</div>
-			</div>
-
-			<br />
-		</div>
-	);
+	return <UserForm fetchUser={fetchUser?.data} refetch={refetch} isLoading={isLoading} />;
 };
 
 export default Profile;
